flaskapp/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-login==0.6.1
Flask-jwt-extended==4.4.1
python-dotenv==0.20.0
pymysql==1.0.2
mariadb==1.0.11
